Solution

+++ app/main.py
@@ -6,7 +6,7 @@ class Car:
 
 
 class CarWashStation:
-    def __init__(self, distance_from_city_center: int, clean_power: int,
+    def __init__(self, distance_from_city_center: float, clean_power: int,
                  average_rating: float, count_of_ratings: int) -> None:
         self.distance_from_city_center = distance_from_city_center
         self.clean_power = clean_power
